+++ changeLog.md
@@ -1,3 +1,17 @@
+# v0.2
+1. fixed jarvis data folder system in linux
+
+2. fixed weather city not found bug
+
+3. added version checker API
+
+4. added password copy to clipboard functionality in password module
+
+5. added more pretty display of tabular data in password module using pandas
+
+6. added table formatting to tabulate in password module
+
+
 # v0.1
 
 1. Designed with modularity in mind that is all the internal function and packages returnes data to main function instead of priting is write away making them reusable anywere